Feat : DTO Fix
Former-commit-id: d32cb84cacb70ada4c535dde69c82c7c7fd36b80

+++ backend/src/main/java/com/a503/churros/service/news/NewsService.java
@@ -9,9 +9,9 @@ public interface NewsService {
     List<Integer> sendRecommend(long userId);
     List<Integer> sendSample();
     void saveReadArti(long userId, int articleId);
-    void recordLike(long userId, long articleId);
+    void recordLike(long userId, int articleId);
     List<Long> getLikeList(long userIdx);
-    void recordDisLike(long userId, long articleId);
+    void recordDisLike(long userId, int articleId);
 
     ArticleDTO getArticleInfo(/*long userId , */long articleId);
     String callNaver(String url);
